requirements.txt
streamlit==1.31.1
pandas==2.2.0
plotly==5.18.0
numpy==1.26.3
python-dotenv==1.0.0
gspread==5.12.0
oauth2client==4.1.3